Move convert_to_tsquery to contracts.models.
